+++ src/main/java/roomescape/common/resolver/LoginMemberArgumentResolver.java
@@ -8,17 +8,17 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import roomescape.member.AuthService;
 import roomescape.member.LoginMember;
 import roomescape.member.Member;
-import roomescape.member.MemberService;
 
 @Component
 public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolver {
 
-    private final MemberService memberService;
+    private final AuthService authService;
 
-    public LoginMemberArgumentResolver(MemberService memberService) {
-        this.memberService = memberService;
+    public LoginMemberArgumentResolver(AuthService authService) {
+        this.authService = authService;
     }
 
     @Override
@@ -32,7 +32,7 @@ public class LoginMemberArgumentResolver implements HandlerMethodArgumentResolve
         Cookie[] cookies = request.getCookies();
         for (Cookie cookie : cookies) {
             if (cookie.getName().equals("token")) {
-                Member member = memberService.loginCheck(cookie.getValue());
+                Member member = authService.loginCheck(cookie.getValue());
                 return new LoginMember(member.getId(), member.getName(), member.getEmail(), member.getRole());
             }
         }
