+++ src/main/java/roomescape/member/MemberService.java
@@ -1,17 +1,14 @@
 package roomescape.member;
 
 import org.springframework.stereotype.Service;
-import roomescape.member.dto.LoginRequest;
 
 @Service
 public class MemberService {
 
     private final MemberDao memberDao;
-    private final TokenProvider tokenProvider;
 
-    public MemberService(MemberDao memberDao, TokenProvider tokenProvider) {
+    public MemberService(MemberDao memberDao) {
         this.memberDao = memberDao;
-        this.tokenProvider = tokenProvider;
     }
 
     public MemberResponse createMember(MemberRequest memberRequest) {
@@ -19,19 +16,4 @@ public class MemberService {
         return new MemberResponse(member.getId(), member.getName(), member.getEmail());
     }
 
-    public String login(LoginRequest loginRequest) {
-        Member findMember = memberDao
-                .findByEmailAndPassword(loginRequest.getEmail(), loginRequest.getPassword())
-                .orElseThrow(() -> new IllegalArgumentException("Invalid email or password"));
-
-        return tokenProvider.createToken(findMember);
-    }
-
-    public Member loginCheck(String token) {
-        Long memberId = tokenProvider.parse(token);
-
-
-        return memberDao.findById(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("Invalid token"));
-    }
 }
